[ALPS05441046] HID: core: Sanitize event and handle ReportSize
1.HID: core: Sanitize event code and type when mapping input
file:hid-core.c  hid-input.c  hid-multitouch.c

When calling into hid_map_usage(), the passed event code is
blindly stored as is, even if it doesnt fit in the associated bitmap.

This event code can come from a variety of sources, including devices
masquerading as input devices, only a bit more "programmable".

Instead of taking the event code at face value, check that it actually
fits the corresponding bitmap, and if it doesnt:
- spit out a warning so that we know which device is acting up
- NULLify the bitmap pointer so that we catch unexpected uses

Code paths that can make use of untrusted inputs can now check
that the mapping was indeed correct and bail out if not.

2.HID: core: Correctly handle ReportSize being zero
file:hid.h

It appears that a ReportSize value of zero is legal, even if a bit
non-sensical. Most of the HID code seems to handle that gracefully,
except when computing the total size in bytes. When fed as input to
memset, this leads to some funky outcomes.

Detect the corner case and correctly compute the size.

MTK-Commit-Id: 3aa8ed27d8aa0ef593a0a6150a0de006ccb7f36c

Change-Id: Ic46dd005dc62819efdc34f6ed9db7077dbd40ff4
Signed-off-by: mtk15754 <wenqiang.zhang@mediatek.com>
CR-Id: ALPS05441046
Feature: Google Security Patch
(cherry-pick from 7f2100811391adc83545348361d6d591260d4e39)